new form component, adding grid
